Add loadScopeData to product resource model
- Will feed data to product edit DataProvider
