+++ app/code/Magento/Eav/Model/ResourceModel/ReadHandler.php
@@ -14,69 +14,19 @@ use Psr\Log\LoggerInterface;
 class ReadHandler implements AttributeInterface
 {
     /**
-     * @var MetadataPool
+     * @var ReadHandler\Attribute
      */
-    protected $metadataPool;
+    protected $attribute;
 
     /**
-     * @var ScopeResolver
-     */
-    protected $scopeResolver;
-
-    /**
-     * @var LoggerInterface
-     */
-    private $logger;
-
-    /** @var \Magento\Eav\Model\Config */
-    private $config;
-
-    /**
-     * ReadHandler constructor.
+     * ReadHandler constructor
      *
-     * @param MetadataPool $metadataPool
-     * @param ScopeResolver $scopeResolver
-     * @param LoggerInterface $logger
-     * @param \Magento\Eav\Model\Config $config
+     * @param ReadHandler\Attribute $attribute
      */
     public function __construct(
-        MetadataPool $metadataPool,
-        ScopeResolver $scopeResolver,
-        LoggerInterface $logger,
-        \Magento\Eav\Model\Config $config
+        \Magento\Eav\Model\ResourceModel\ReadHandler\Attribute $attribute
     ) {
-        $this->metadataPool = $metadataPool;
-        $this->scopeResolver = $scopeResolver;
-        $this->logger = $logger;
-        $this->config = $config;
-    }
-
-    /**
-     * Get attribute of given entity type
-     *
-     * @param string $entityType
-     * @return \Magento\Eav\Api\Data\AttributeInterface[]
-     * @throws \Exception if for unknown entity type
-     */
-    protected function getAttributes($entityType)
-    {
-        $metadata = $this->metadataPool->getMetadata($entityType);
-        $eavEntityType = $metadata->getEavEntityType();
-        $attributes = (null === $eavEntityType) ? [] : $this->config->getAttributes($eavEntityType);
-        return $attributes;
-    }
-
-    /**
-     * @param ScopeInterface $scope
-     * @return array
-     */
-    protected function getContextVariables(ScopeInterface $scope)
-    {
-        $data[] = $scope->getValue();
-        if ($scope->getFallback()) {
-            $data = array_merge($data, $this->getContextVariables($scope->getFallback()));
-        }
-        return $data;
+        $this->attribute = $attribute;
     }
 
     /**
@@ -93,105 +43,6 @@ class ReadHandler implements AttributeInterface
      */
     public function execute($entityType, $entityData, $arguments = [])
     {
-        return $this->getData($entityType, $entityData, $arguments = []);
-    }
-
-    /**
-     * @param string $entityType
-     * @param array $entityData
-     * @param array $arguments
-     * @return array
-     * @throws \Exception
-     * @throws \Magento\Framework\Exception\ConfigurationMismatchException
-     * @throws \Magento\Framework\Exception\LocalizedException
-     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
-     */
-    public function getAllScopeData($entityType, $entityData, $arguments = [])
-    {
-        return $this->getData($entityType, $entityData, $arguments = [], true);
-    }
-
-    /**
-     * @param string $entityType
-     * @param array $entityData
-     * @param array $arguments
-     * @param bool $loadAllScopes
-     * @return array
-     * @throws \Exception
-     * @throws \Magento\Framework\Exception\ConfigurationMismatchException
-     * @throws \Magento\Framework\Exception\LocalizedException
-     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
-     */
-    public function getData($entityType, $entityData, $arguments = [], $loadAllScopes = false)
-    {
-        $metadata = $this->metadataPool->getMetadata($entityType);
-        if (!$metadata->getEavEntityType()) {//todo hasCustomAttributes
-            return $entityData;
-        }
-        $context = $this->scopeResolver->getEntityContext($entityType, $entityData);
-        $connection = $metadata->getEntityConnection();
-
-        $attributeTables = [];
-        $attributesMap = [];
-        $selects = [];
-
-        /** @var \Magento\Eav\Model\Entity\Attribute\AbstractAttribute $attribute */
-        foreach ($this->getAttributes($entityType) as $attribute) {
-            if (!$attribute->isStatic()) {
-                $attributeTables[$attribute->getBackend()->getTable()][] = $attribute->getAttributeId();
-                $attributesMap[$attribute->getAttributeId()] = $attribute->getAttributeCode();
-            }
-        }
-        if (count($attributeTables)) {
-            $attributeTables = array_keys($attributeTables);
-            foreach ($attributeTables as $attributeTable) {
-                $from = ['value' => 't.value', 'attribute_id' => 't.attribute_id'];
-                if ($loadAllScopes) {
-                    $from['store_id'] = 't.store_id';
-                }
-                $select = $connection->select()
-                    ->from(
-                        ['t' => $attributeTable],
-                        $from
-                    )
-                    ->where($metadata->getLinkField() . ' = ?', $entityData[$metadata->getLinkField()]);
-                // TODO: Clean up this code
-                if (!$loadAllScopes) {
-                    foreach ($context as $scope) {
-                        //TODO: if (in table exists context field)
-                        $select->where(
-                            $metadata->getEntityConnection()->quoteIdentifier($scope->getIdentifier()) . ' IN (?)',
-                            $this->getContextVariables($scope)
-                        )->order('t.' . $scope->getIdentifier() . ' DESC');
-                    }
-                }
-                $selects[] = $select;
-            }
-            $unionSelect = new \Magento\Framework\DB\Sql\UnionExpression(
-                $selects,
-                \Magento\Framework\DB\Select::SQL_UNION_ALL
-            );
-
-            foreach ($connection->fetchAll($unionSelect) as $attributeValue) {
-                if (isset($attributesMap[$attributeValue['attribute_id']])) {
-                    if ($loadAllScopes) {
-                        $scopeData[$attributesMap[$attributeValue['attribute_id']]][$attributeValue['store_id']] = $attributeValue['value'];
-                    } else {
-                        $entityData[$attributesMap[$attributeValue['attribute_id']]] = $attributeValue['value'];
-                    }
-                } else {
-                    $this->getLogger()->warning(
-                        "Attempt to load value of nonexistent EAV attribute '{$attributeValue['attribute_id']}'
-                        for entity type '$entityType'."
-                    );
-                }
-            }
-        }
-
-        // TODO: Clean up this code
-        if ($loadAllScopes) {
-            return $scopeData;
-        }
-        return $entityData;
+        return $this->attribute->getData($entityType, $entityData, $arguments = []);
     }
 }
